+++ app/Models/UserLedgerEntry.php
@@ -84,8 +84,8 @@ class UserLedgerEntry extends Model
     public const TYPE_BONUS_INCOME = 5;
     public const TYPE_BONUS_EXPENSE = 6;
     public const TYPE_REFUND = 7;
-    public const TYPE_NON_REFUNDABLE_DEPOSIT = 8;
-    public const TYPE_NON_REFUNDABLE_EXPENSE = 9;
+    public const TYPE_NON_WITHDRAWABLE_DEPOSIT = 8;
+    public const TYPE_NON_WITHDRAWABLE_EXPENSE = 9;
 
     public const ALLOWED_STATUS_LIST = [
         self::STATUS_ACCEPTED,
@@ -108,8 +108,8 @@ class UserLedgerEntry extends Model
         self::TYPE_BONUS_INCOME,
         self::TYPE_BONUS_EXPENSE,
         self::TYPE_REFUND,
-        self::TYPE_NON_REFUNDABLE_DEPOSIT,
-        self::TYPE_NON_REFUNDABLE_EXPENSE,
+        self::TYPE_NON_WITHDRAWABLE_DEPOSIT,
+        self::TYPE_NON_WITHDRAWABLE_EXPENSE,
     ];
 
     public const CREDIT_TYPES = [
@@ -117,14 +117,14 @@ class UserLedgerEntry extends Model
         self::TYPE_REFUND,
         self::TYPE_AD_INCOME,
         self::TYPE_BONUS_INCOME,
-        self::TYPE_NON_REFUNDABLE_DEPOSIT,
+        self::TYPE_NON_WITHDRAWABLE_DEPOSIT,
     ];
 
     public const DEBIT_TYPES = [
         self::TYPE_WITHDRAWAL,
         self::TYPE_AD_EXPENSE,
         self::TYPE_BONUS_EXPENSE,
-        self::TYPE_NON_REFUNDABLE_EXPENSE,
+        self::TYPE_NON_WITHDRAWABLE_EXPENSE,
     ];
 
     private const AWAITING_PAYMENTS = [
@@ -192,8 +192,8 @@ class UserLedgerEntry extends Model
             ->whereNotIn('type', [
                 self::TYPE_BONUS_INCOME,
                 self::TYPE_BONUS_EXPENSE,
-                self::TYPE_NON_REFUNDABLE_DEPOSIT,
-                self::TYPE_NON_REFUNDABLE_EXPENSE
+                self::TYPE_NON_WITHDRAWABLE_DEPOSIT,
+                self::TYPE_NON_WITHDRAWABLE_EXPENSE
             ]);
     }
 
@@ -313,7 +313,7 @@ class UserLedgerEntry extends Model
     public static function removeProcessingExpenses(): void
     {
         self::where('status', self::STATUS_PROCESSING)
-            ->whereIn('type', [self::TYPE_AD_EXPENSE, self::TYPE_NON_REFUNDABLE_EXPENSE, self::TYPE_BONUS_EXPENSE])
+            ->whereIn('type', [self::TYPE_AD_EXPENSE, self::TYPE_NON_WITHDRAWABLE_EXPENSE, self::TYPE_BONUS_EXPENSE])
             ->delete();
     }
 
@@ -334,7 +334,7 @@ class UserLedgerEntry extends Model
     private static function blockedEntries(): Builder
     {
         return self::where('status', self::STATUS_BLOCKED)
-            ->whereIn('type', [self::TYPE_AD_EXPENSE, self::TYPE_NON_REFUNDABLE_EXPENSE, self::TYPE_BONUS_EXPENSE]);
+            ->whereIn('type', [self::TYPE_AD_EXPENSE, self::TYPE_NON_WITHDRAWABLE_EXPENSE, self::TYPE_BONUS_EXPENSE]);
     }
 
     private static function blockedEntriesByUserId(int $userId): Builder
@@ -352,16 +352,16 @@ class UserLedgerEntry extends Model
 
         $user = User::findOrFail($userId);
         $bonusableAmount = (int)max(min($total, $maxBonus, $user->getBonusBalance()), 0);
-        $nonRefundableAmount = (int)max(
+        $nonWithdrawableAmount = (int)max(
             min($total - $bonusableAmount, $user->getWalletBalance() - $user->getWithdrawableBalance()),
             0
         );
-        $refundableAmount = (int)max(
-            min($total - $nonRefundableAmount - $bonusableAmount, $user->getWalletBalance()),
+        $withdrawableAmount = (int)max(
+            min($total - $nonWithdrawableAmount - $bonusableAmount, $user->getWalletBalance()),
             0
         );
 
-        if ($total > $bonusableAmount + $nonRefundableAmount + $refundableAmount) {
+        if ($total > $bonusableAmount + $nonWithdrawableAmount + $withdrawableAmount) {
             throw new InvalidArgumentException(
                 sprintf('Insufficient funds for User [%s] when adding ad expense.', $userId)
             );
@@ -369,8 +369,8 @@ class UserLedgerEntry extends Model
 
         $entries = [
             self::insertAdExpense($status, $userId, $bonusableAmount, self::TYPE_BONUS_EXPENSE),
-            self::insertAdExpense($status, $userId, $nonRefundableAmount, self::TYPE_NON_REFUNDABLE_EXPENSE),
-            self::insertAdExpense($status, $userId, $refundableAmount, self::TYPE_AD_EXPENSE),
+            self::insertAdExpense($status, $userId, $nonWithdrawableAmount, self::TYPE_NON_WITHDRAWABLE_EXPENSE),
+            self::insertAdExpense($status, $userId, $withdrawableAmount, self::TYPE_AD_EXPENSE),
         ];
 
         if (
@@ -381,12 +381,12 @@ class UserLedgerEntry extends Model
         ) {
             $entries[] = self::insertUserRefund(
                 $refLink->user_id,
-                $refLink->calculateRefund($nonRefundableAmount + $refundableAmount),
+                $refLink->calculateRefund($nonWithdrawableAmount + $withdrawableAmount),
                 $refLink
             );
             $entries[] = self::insertUserBonus(
                 $userId,
-                $refLink->calculateBonus($nonRefundableAmount + $refundableAmount),
+                $refLink->calculateBonus($nonWithdrawableAmount + $withdrawableAmount),
                 $refLink
             );
         }